Remove some unused. Move some stuff around.
